dojson: moved receiver to record receivers
