+++ inspirehep/modules/records/receivers.py
@@ -27,6 +27,7 @@ from __future__ import absolute_import, division, print_function
 import six
 from flask import current_app
 from flask_sqlalchemy import models_committed
+from itertools import chain
 
 from invenio_indexer.api import RecordIndexer
 from invenio_indexer.signals import before_record_index
@@ -35,7 +36,7 @@ from invenio_records.signals import before_record_insert, before_record_update
 
 from inspirehep.dojson.utils import classify_field, get_recid_from_ref
 from inspirehep.modules.records.api import InspireRecord
-from inspirehep.utils.date import create_valid_date
+from inspirehep.utils.date import create_valid_date, create_earliest_date
 from inspirehep.utils.helpers import force_force_list
 from inspirehep.utils.record import get_value
 
@@ -65,7 +66,6 @@ def enhance_record(sender, json, *args, **kwargs):
     dates_validator(sender, json, *args, **kwargs)
     add_recids_and_validate(sender, json, *args, **kwargs)
     populate_experiment_suggest(sender, json, *args, **kwargs)
-    populate_abstract_source_suggest(sender, json, *args, **kwargs)
     after_record_enhanced.send(json)
 
 
@@ -397,3 +397,23 @@ def populate_abstract_source_suggest(sender, json, *args, **kwargs):
                         'output': source,
                     },
                 })
+
+
+@before_record_index.connect
+def earliest_date(sender, json, *args, **kwargs):
+    """Find and assign the earliest date to a HEP paper."""
+    date_paths = [
+        'preprint_date',
+        'thesis.date',
+        'thesis.defense_date',
+        'publication_info.year',
+        'creation_modification_date.creation_date',
+        'imprints.date',
+    ]
+
+    dates = list(chain.from_iterable(
+        [force_force_list(get_value(json, path)) for path in date_paths]))
+
+    earliest_date = create_earliest_date(dates)
+    if earliest_date:
+        json['earliest_date'] = earliest_date
